chore: moved docker to pgk directory
