+++ automated-testing/pkg/docker/images.go
